requirements.txt
Django~=4.2.1
WeasyPrint~=52.4
DateTime~=5.5
djangorestframework~=3.15.2